LiveSync: move BlockQueue to own task
